edit user fixed and added personal tutor feature for student

+++ server/controller/teachers.js
@@ -4,6 +4,7 @@
 
 const Module = require('../model/Module.js');
 const Teacher = require('../model/Teacher.js');
+const mongoose = require('mongoose');
 
 exports.find = async (_id) => {
 	if (!_id) {
@@ -11,18 +12,29 @@ exports.find = async (_id) => {
 
 		return teachers;
 	}
-	const teachers = await Teacher.findById(_id).populate('user').populate('module');
+	if (!mongoose.isValidObjectId(_id)) {
+		return { err: "Teacher doesnot exist" };
+	}
+	const teachers = await Teacher.findById(_id, (err, data) => { if (err) { console.log('err') } }).populate('user').populate('module');
 	return teachers;
 }
 
 exports.findByModule = async (module) => {
+	if (!mongoose.isValidObjectId(module)) {
+		return { err: "Module doesnot exist" };
+	}
 	const teachers = await Teacher.find({ module }).populate('user').populate('module');
-	return teachers
+	console.log(teachers);
+	return teachers;
 }
 
 
 exports.getAvailableTeachers = async (module) => {
-	let teachers = await Teacher.find({ module }).populate('user').populate('module');
+	if (!mongoose.isValidObjectId(module)) {
+		return { err: "Module doesnot exist" };
+	}
+	let teachers = await Teacher.find({ module: module }).populate('user').populate('module');
+	console.log(teachers);
 	for (let i in teachers) {
 		if (teachers.personalstudentId != undefined) {
 			teachers.splice(i);
@@ -32,7 +44,10 @@ exports.getAvailableTeachers = async (module) => {
 }
 
 exports.removeModule = async (module) => {
-	const teachers = await Teacher.find({ module });
+	if (!mongoose.isValidObjectId(module)) {
+		return { err: "Module doesnot exist" };
+	}
+	const teachers = await Teacher.find({ module }).populate();
 	for (let i = 0; i < teachers.length; i++) {
 		await Teacher.findByIdAndUpdate(teachers[i]._id, { module: undefined });
 	}
@@ -40,26 +55,65 @@ exports.removeModule = async (module) => {
 
 exports.add = async (user) => {
 
+
 	const { name, email, dob, phoneNumber, address, role, module } = user;
 	const User = await require('../controller/users.js').add({ name, email, dob, phoneNumber, address, role });
+	let teacher;
+	if (role == "admin" || module == "none") {
+		teacher = new Teacher({
+			user: User._id,
+		})
+	}
+	else {
+		if (!mongoose.isValidObjectId(module)) {
+			return { err: "Module doesnot exist" };
+		}
 
-	const teacher = new Teacher({
-		user: User._id,
-		module
+		teacher = new Teacher({
+			user: User._id,
+			module
 
-	})
+		})
+	}
 	await teacher.save();
+	return {};
 }
 
 exports.edit = async (edited) => {
 
-	const { name, email, dob, phoneNumber, address, salary, module } = edited;
-	const User = await require('../controller/users.js').edit({ name, email, dob, phoneNumber, address, role: "student" });
-	const teacher = await teacher.findByIdAndUpdate({ _id }, {
+	const { name, email, dob, phoneNumber, address, salary, module, _id, role } = edited;
+	if (!mongoose.isValidObjectId(_id)) {
+		console.log('here');
+		return { err: "Teacher does not exist" };
+	}
+
+
+	const teacher = await Teacher.findById(_id).populate('user');
+	if (!teacher) {
+		return { err: "Teacher does not exist" };
+	}
+
+
+	await require('../controller/users').edit({ name, email, dob, phoneNumber, address, role }, teacher.user._id);
+
+	if (role == "admin" || module == "none") {
+		await Teacher.findByIdAndUpdate(_id, {
+			salary
+		});
+	}
+
+
+	if (!mongoose.isValidObjectId(module)) {
+		return { err: "Module id Error" };
+	}
+
+
+	await Teacher.findByIdAndUpdate(_id, {
 		salary,
 		module
 	});
 
+	return {};
 }
 
 exports.getID = async (uid) => {
@@ -68,6 +122,9 @@ exports.getID = async (uid) => {
 }
 
 exports.delete = async (_id) => {
+	if (!mongoose.isValidObjectId(_id)) {
+		return { err: "notValidObjectId" };
+	}
 	let teacher = await Teacher.findById({ _id }).populate('user');
 	await require('./users').delete(teacher.user._id);
 	await Teacher.deleteOne({ _id });
